Swap Index Early (#1353)
